refactor: replace DoctypeModal with DoctypeModals component for better modularity
