+++ frontend/src/App.vue
@@ -5,31 +5,19 @@
       <router-view :key="$route.fullPath" />
     </Layout>
     <Dialogs />
+    <DoctypeModals />
     <EventNotificationPopup />
-    <DoctypeModal
-      v-if="doctypeModal.show.value"
-      v-model="doctypeModal.show.value"
-      :doctypeTitle="doctypeModal.title.value"
-      :doctype="doctypeModal.doctype.value"
-      :docname="doctypeModal.name.value"
-      :defaults="doctypeModal.defaults.value"
-      @afterInsert="(d) => doctypeModal.triggerCallback('afterInsert', d)"
-      @afterUpdate="(d) => doctypeModal.triggerCallback('afterUpdate', d)"
-    />
   </FrappeUIProvider>
 </template>
 
 <script setup>
 import NotPermitted from '@/pages/NotPermitted.vue'
 import EventNotificationPopup from '@/components/EventNotificationPopup.vue'
-import DoctypeModal from '@/components/Modals/DoctypeModal.vue'
+import DoctypeModals from '@/components/Modals/DoctypeModals.vue'
 import { Dialogs } from '@/utils/dialogs'
-import { useDoctypeModal } from '@/composables/doctypeModal'
 import { sessionStore } from '@/stores/session'
 import { FrappeUIProvider, setConfig, useTheme } from 'frappe-ui'
 import { computed, defineAsyncComponent, provide } from 'vue'
-
-const doctypeModal = useDoctypeModal()
 
 const session = sessionStore()
 provide('session', session)
