+++ frontend/src/components/Modals/DoctypeModals.vue
@@ -1,0 +1,18 @@
+<template>
+  <DoctypeModal
+    v-if="doctypeModal.show.value"
+    v-model="doctypeModal.show.value"
+    :doctypeTitle="doctypeModal.title.value"
+    :doctype="doctypeModal.doctype.value"
+    :docname="doctypeModal.name.value"
+    :defaults="doctypeModal.defaults.value"
+    @afterInsert="(d) => doctypeModal.triggerCallback('afterInsert', d)"
+    @afterUpdate="(d) => doctypeModal.triggerCallback('afterUpdate', d)"
+  />
+</template>
+<script setup>
+import DoctypeModal from '@/components/Modals/DoctypeModal.vue'
+import { useDoctypeModal } from '@/composables/doctypeModal'
+
+const doctypeModal = useDoctypeModal()
+</script>
